Separate the description text flow container's test case.
